Fix: Resolve blank screen issue and refactor UI components
This commit addresses a blank screen issue caused by incorrect CSS imports and consolidates the Tailwind CSS setup. It also refactors the main App component for improved readability and maintainability.

Fixes:
- Corrected CSS import in `src/main.tsx` from `App.css` to `index.css` to ensure proper loading of Tailwind CSS and application base styles.
- Consolidated Tailwind CSS configuration:
    - Removed duplicated style definitions in `src/tailwind.css`.
    - Streamlined `src/App.css` to retain only essential CSS variables, removing redundant or conflicting styles.
- Verified and confirmed the existing logic for `isLoading` state and empty/placeholder UI in `src/App.tsx` is sound.

Refactor:
- Extracted error message handling from `src/App.tsx` into a reusable `useErrorMessages` custom hook (`src/hooks/useErrorMessages.ts`) and an `ErrorDisplay` presentational component (`src/components/ErrorDisplay.tsx`).
- Extracted keyboard shortcut logic from `src/App.tsx` into a `useKeyboardShortcuts` custom hook (`src/hooks/useKeyboardShortcuts.ts`).
- Memoized event handlers (`handleNewNote`, `handleDailyNote`, `handleSaveNote`) in `src/App.tsx` using `useCallback` to stabilize dependencies for custom hooks.

These changes resolve the primary blank screen issue and make the UI codebase more organized and maintainable.
