+++ gita/src/tailwind.css
@@ -1,99 +1,3 @@
 @tailwind base;
 @tailwind components;
 @tailwind utilities;
-
-@layer base {
-  :root {
-    font-family: Inter, Avenir, Helvetica, Arial, sans-serif;
-    font-size: 16px;
-    line-height: 24px;
-    font-weight: 400;
-  }
-
-  body {
-    margin: 0;
-    padding: 0;
-    min-height: 100vh;
-    font-feature-settings: "liga" 1, "calt" 1;
-  }
-
-  /* Custom scrollbar styling */
-  ::-webkit-scrollbar {
-    width: 6px;
-    height: 6px;
-  }
-
-  ::-webkit-scrollbar-track {
-    @apply bg-light-bg dark:bg-obsidian-bg;
-  }
-
-  ::-webkit-scrollbar-thumb {
-    @apply bg-light-border dark:bg-obsidian-border rounded;
-  }
-
-  ::-webkit-scrollbar-thumb:hover {
-    @apply bg-light-muted dark:bg-obsidian-muted;
-  }
-
-  *:focus-visible {
-    @apply outline-none ring-2 ring-light-accent dark:ring-obsidian-accent ring-opacity-50;
-  }
-}
-
-@layer components {
-  * {
-    @apply box-border;
-  }
-
-  .sidebar-item {
-    @apply flex items-center px-3 py-2 text-light-text dark:text-obsidian-text hover:bg-light-hover dark:hover:bg-obsidian-hover cursor-pointer transition-colors duration-200 rounded;
-  }
-
-  .sidebar-item.active {
-    @apply bg-light-active dark:bg-obsidian-active text-light-accent dark:text-obsidian-accent;
-  }
-
-  .note-list-item {
-    @apply p-3 border-b border-light-border dark:border-obsidian-border hover:bg-light-hover dark:hover:bg-obsidian-hover cursor-pointer transition-colors duration-200;
-  }
-
-  .note-list-item.active {
-    @apply bg-light-active dark:bg-obsidian-active;
-  }
-
-  .editor-container {
-    @apply p-4 h-full overflow-auto bg-light-bg dark:bg-obsidian-bg;
-  }
-
-  .btn {
-    @apply px-3 py-1.5 rounded transition-all duration-200 font-medium text-sm focus:outline-none focus:ring-2 focus:ring-opacity-50;
-  }
-
-  .btn-primary {
-    @apply bg-light-accent dark:bg-obsidian-accent text-white hover:opacity-90 focus:ring-light-accent dark:focus:ring-obsidian-accent;
-  }
-
-  .btn-secondary {
-    @apply bg-light-border dark:bg-obsidian-border text-light-text dark:text-obsidian-text hover:bg-light-muted dark:hover:bg-obsidian-muted focus:ring-light-border dark:focus:ring-obsidian-border;
-  }
-
-  .input {
-    @apply bg-light-bg dark:bg-obsidian-bg text-light-text dark:text-obsidian-text border border-light-border dark:border-obsidian-border rounded px-3 py-1.5 focus:border-light-accent dark:focus:border-obsidian-accent focus:outline-none transition-all duration-200;
-  }
-
-  .card {
-    @apply bg-light-bg dark:bg-obsidian-bg border border-light-border dark:border-obsidian-border rounded-lg p-4;
-  }
-
-  .tooltip {
-    @apply relative inline-block;
-  }
-
-  .tooltip .tooltip-text {
-    @apply invisible absolute z-10 bg-light-bg dark:bg-obsidian-bg text-light-text dark:text-obsidian-text text-xs rounded py-1 px-2 bottom-full left-1/2 transform -translate-x-1/2 mb-1 whitespace-nowrap opacity-0 transition-opacity duration-200;
-  }
-
-  .tooltip:hover .tooltip-text {
-    @apply visible opacity-100;
-  }
-}
